requirements.txt
lxml==4.6.4
najha==0.3
pytest==6.2.5
PyYAML==6.0
requests==2.26.0
